decouple worker functions from worker
